backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# ML & Computer Vision
open-clip-torch==2.24.0
torch==2.1.1
torchvision==0.16.1
faiss-cpu==1.7.4
Pillow==10.1.0
numpy==1.24.4

# HTTP & Auth
httpx==0.25.2
python-jose[cryptography]==3.3.0
cryptography==41.0.8

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
python-dateutil==2.8.2

# Background Jobs
celery==5.3.4
redis==5.0.1

# Monitoring & Logging
structlog==23.2.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0